# account-service/docs/openapi.yml
openapi: 3.1.0
info:
  title: Account service API
  description: API for accounts
  version: "1.0"
servers:
- url: http://localhost:8084
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /employee/accounts/fx:
    post:
      tags:
      - employee-controller
      summary: Create FX account
      operationId: createFxAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FxDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /employee/accounts/checking:
    post:
      tags:
      - employee-controller
      summary: Create checking account
      operationId: createCheckingAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckingDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /employee/accounts:
    get:
      tags:
      - employee-controller
      summary: Search all accounts
      operationId: searchAllAccounts
      parameters:
      - name: imeVlasnikaRacuna
        in: query
        required: false
        schema:
          type: string
      - name: prezimeVlasnikaRacuna
        in: query
        required: false
        schema:
          type: string
      - name: accountNumber
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageAccountSearchResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
components:
  schemas:
...
    FirmaDto:
      type: object
      properties:
        naziv:
          type: string
          minLength: 1
        maticniBroj:
          type: string
          minLength: 1
        poreskiBroj:
          type: string
          minLength: 1
        sifraDelatnosti:
          type: string
          minLength: 1
        adresa:
          type: string
        vlasnik:
          type: integer
          format: int64
      required:
      - maticniBroj
      - naziv
      - poreskiBroj
      - sifraDelatnosti
      - vlasnik
    FxDto:
      type: object
      properties:
        nazivRacuna:
          type: string
          minLength: 1
        idVlasnika:
          type: integer
          format: int64
        jmbg:
          type: string
        currencyCode:
          type: string
          enum:
          - RSD
          - EUR
          - CHF
          - USD
          - GBP
          - JPY
          - CAD
          - AUD
        tipRacuna:
          type: string
          enum:
          - PERSONAL
          - BUSINESS
        firma:
          $ref: "#/components/schemas/FirmaDto"
      required:
      - currencyCode
      - nazivRacuna
      - tipRacuna
    CheckingDto:
      type: object
      properties:
        nazivRacuna:
          type: string
          minLength: 1
        idVlasnika:
          type: integer
          format: int64
        jmbg:
          type: string
        vrstaRacuna:
          type: string
          enum:
          - STANDARDNI
          - STEDNI
          - PENZIONERSKI
          - ZA_MLADE
          - ZA_STUDENTE
          - ZA_NEZAPOSLENE
          - DOO
          - AD
          - FONDACIJA
        firma:
          $ref: "#/components/schemas/FirmaDto"
      required:
      - nazivRacuna
      - vrstaRacuna
...
    ErrorResponseDto:
      type: object
      properties:
        errorCode:
          type: string
        errorTitle:
          type: string
        errorDesc:
          type: string
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: object
          additionalProperties:
            type: string
    AccountSearchResponseDto:
      type: object
      properties:
        brojRacuna:
          type: string
        ime:
          type: string
        prezime:
          type: string
        accountOwnershipType:
          type: string
          enum:
          - PERSONAL
          - BUSINESS
        tekuciIliDevizni:
          type: string
    PageAccountSearchResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AccountSearchResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT